% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_mean_ci.R
\name{accuracy_mean_ci}
\alias{accuracy_mean_ci}
\title{Calculate average accuracy and its CI}
\usage{
accuracy_mean_ci(blrf, data, lower = 0.025, upper = 0.975)
}
\arguments{
\item{blrf}{Object of class inheriting from \code{blrf}.}

\item{data}{data.frame object. Data to be predicted for accuracy.}

\item{lower}{numeric. Define lower bound of quantile.
Default to be 0.025.}

\item{upper}{numeric. Define upper bound of quantile.
Default to be 0.975.}
}
\value{
matrix. Average accuracy and confidence interval for each response variable.
}
\description{
Function is used to compute average accuracy given input of "factor" type of \code{blrf}
and its the confidence interval of accuracy aggreating all Trees.
Will call Confusion_one_tree function to calculate confusion matrix for one tree.
Default to output 95 percentage confidence interval.
}
\examples{
tt <- blrf(Species~., iris, gamma = 0.7, s = 10, r = 100, n_var = 2)
accuracy_mean_ci(tt, data, lower = 0.025, upper = 0.975)
}
