% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrf.R
\name{blrf}
\alias{blrf}
\title{Fitting random forest with bag of little bootstraps}
\usage{
blrf(
  formula,
  data,
  gamma,
  b = NULL,
  s,
  r,
  n_var,
  split = "gini",
  control = tree::tree.control(nobs = nrow(data), minsize = 10),
  core = 1
)
}
\arguments{
\item{formula}{an object of class \code{formula}.}

\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{gamma}{numeric. User_defined sizing factor.}

\item{b}{numeric. An optional number of size of subsamples.}

\item{s}{numeric. number of subsamples.}

\item{r}{numeric. Number of trees.}

\item{n_var}{numeric. Number of variables to randomly subset to build one tree.}

\item{split}{character string. Can be "deviance" or "gini". Default to be "gini".}

\item{control}{control object based on tree::tree.control() function from "tree" package.
Default set to call tree::tree.control(), ie, using default values from tree::tree.control()).}

\item{core}{numeric. Must be positive. Number of core to use for parallel computing.
Default to 1, meaning no use of parallel computing.
If higher than 1, then implement the function with parallel computing with given number of cores.}
}
\value{
Object of class inheriting from \code{blrf}.
}
\description{
blrf is used to fit Random Forest with Bag of Little Random Forests.
It can be used to carry out both regression and classification trees.
}
\details{
\code{gamma} is used to indicate the factor of the size to subsample the original data,
ie. \eqn{n^{gamma}}.
Optional \code{b} indicates the size of each subsample and each sample are set to be equal sized if b is given.
\code{n_var} is the number of variables to be used in the tree and
is used to randomly select variables when each tree is generated.
If \code{b} and \code{gamma} are both given, \code{b} is used to generate subsamples.
\code{split} can be "gini" or "deviance" and is the splitting criterion to use to generate each tree.
Detail argument of \code{control} can be found in tree::tree.control().
}
\examples{
data(iris)
blrf(Species~., iris, gamma = 0.7, s = 10, r = 100, n_var = 2)
}
