% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_check_input.R
\name{implement_check_input}
\alias{implement_check_input}
\title{Checking if all inputs are valid for function implement_BLRF.
If not valid input exists, terminate program and output error message.}
\usage{
implement_check_input(x, y, formula, data, gamma, b, s, r, n_var, split, core)
}
\arguments{
\item{x}{character vector. vector of x variables' names.}

\item{y}{character. y variable name.}

\item{formula}{A formula expression.}

\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{gamma}{numeric. User_defined sizing factor.}

\item{b}{numeric. Size of subsamples.}

\item{s}{numeric. Number of subsamples.}

\item{r}{numeric. Number of trees.}

\item{n_var}{numeric. Number of variables to subset to build one tree.}

\item{split}{character string. Need to be "deviance" or "gini".}

\item{core}{numeric. Must be positive. Number of core to use for parallel computing.}
}
\value{
logic. TRUE if no error, otherwise terminate the program.
}
\description{
Checking if all inputs are valid for function implement_BLRF.
If not valid input exists, terminate program and output error message.
}
