% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_tree.R
\name{one_tree}
\alias{one_tree}
\title{Implement random forest once}
\usage{
one_tree(formula, subsample, weights, n_var, split, control)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{subsample}{data.frame.}

\item{weights}{list.}

\item{n_var}{numeric. Number of variables to randomly subset to build one tree.}

\item{split}{character string. Can be "deviance" or "gini".}

\item{control}{control object based on tree::tree.control() function from "tree" package.}
}
\value{
tree object.
}
\description{
Implement random forest once
}
