% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_one_tree.R
\name{plot_one_tree}
\alias{plot_one_tree}
\title{Plot of one tree}
\usage{
plot_one_tree(
  Tree,
  splits = TRUE,
  label = "yval",
  all = TRUE,
  pretty = 1,
  cex = 1
)
}
\arguments{
\item{Tree}{tree object.}

\item{splits}{logical. If TRUE the splits are labelled.}

\item{label}{The name of column in the frame component of x.
to be used to label the nodes. Can be NULL to suppress node-labelling.}

\item{all}{logical. By default, splits of tree are labeled, ie. interior nodes are also labelled.}

\item{pretty}{the manipulation used for split labels involving attributes.}

\item{cex}{numeric. NULL and NA are equivalent to 1.0.}
}
\value{
plot object.
}
\description{
Plot one tree with given text object of plot.
}
\details{
Options' deatils except `Tree` can be found in tree::text.tree.
}
