% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.blrf.R
\name{predict.blrf}
\alias{predict.blrf}
\title{Predict with given data.}
\usage{
\method{predict}{blrf}(
  blrf,
  newdata,
  confidence = F,
  probability = F,
  pretty = F,
  lower = 0.025,
  upper = 0.975
)
}
\arguments{
\item{blrf}{Object of class inheriting from \code{blrf}.}

\item{newdata}{data.frame. Data to be predicted.}

\item{confidence}{logical. If TRUE, then output will include confidence interval.}

\item{probability}{logical. If TRUE, then output will be predict probability for "factor"
type of \code{blrf}. If FALSE, then the output will be predict label for "factor"
type of \code{blrf} or predict value for "numeric" type of \code{blrf}.}

\item{pretty}{logical. If pretty is TRUE, then output character string output of ci for factor response,
not available for numeric response.}

\item{lower}{numeric. If confidence is TRUE, then define lower bound of ci.
Default to be 0.025.}

\item{upper}{numeric. If confidence is TRUE, then define upper bound of ci.
Default to be 0.975.}
}
\value{
list or matrix of prediction values (or with confidence interval).
}
\description{
User can indicate if output needs to be probability
for factor response or include confidence interval.
Default to output 95% confidence interval if include confidence interval.
}
\details{
\code{confidence} is not available when input is "factor" type of \code{blrf} and
\code{probability} is FALSE.
\code{probability} will not affect output when input is "numeric" type of \code{blrf},
the output will still be the predicted value of averaging all regression trees.
If \code{pretty} is TRUE, then output will contain character string type of confidence interval,
make it more readable. The default range of confidence interval is 95%.
}
\examples{
data(iris)
tt <- blrf(Species~., iris, gamma = 0.7, s = 10, r = 100, n_var = 2)
predict.blrf(tt, iris)
}
