% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_trees.R
\name{print_trees}
\alias{print_trees}
\title{Plot of trees}
\usage{
print_trees(
  blrf,
  tree_list,
  splits = TRUE,
  label = "yval",
  all = TRUE,
  pretty = 1,
  cex = 0.8
)
}
\arguments{
\item{blrf}{Object of class inheriting from \code{blrf}.}

\item{tree_list}{numeric. indext of trees needs to make plot, ie. c(1,2).}

\item{splits}{logical. If TRUE the splits are labelled.}

\item{label}{The name of column in the frame component of x.
to be used to label the nodes. Can be NULL to suppress node-labelling.}

\item{all}{logical. By default, splits of tree are labeled, ie. interior nodes are also labelled.}

\item{pretty}{the manipulation used for split labels involving attributes.}

\item{cex}{numeric. NULL and NA are equivalent to 1.0.}
}
\value{
list of plots.
}
\description{
Plot multiple trees with given the list/vector indexes of target trees in the blrf object.
}
\details{
Options' deatils except \code{blrf} can be found in tree::text.tree().
}
\examples{
data(iris)
tt <- blrf(Species~., iris, gamma = 0.7, s = 10, r = 100, n_var = 2)
}
