% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_ci.R
\name{residual_ci}
\alias{residual_ci}
\title{Calculate average accuracy and its CI}
\usage{
residual_ci(blrf, data, lower = 0.025, upper = 0.975)
}
\arguments{
\item{blrf}{Object of class inheriting from \code{blrf}.}

\item{data}{data.frame.}

\item{lower}{numeric. Define lower bound of residual confidence interval.
Default to be 0.025.}

\item{upper}{numeric. Define upper bound of residual confidence interval.
Default to be 0.975.}
}
\value{
matrix of residual confidence intervals for given observations.
}
\description{
Function is used to compute average residual given input of "numeric" type of \code{blrf}
and its the confidence interval of residuals aggreating all Trees. Default to output 95 percentage
confidence interval.
}
\examples{
data(mtcars)
tt <- blrf(mpg~., mtcars, gamma = 0.7, s = 10, r = 100, n_var = 2)
predict.blrf(tt, mtcars)
}
