% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling.R
\name{subsampling}
\alias{subsampling}
\title{Sampling original data into s subsamples without replacement}
\usage{
subsampling(data, gamma, b = NULL, s)
}
\arguments{
\item{data}{data.frame. Original data.}

\item{gamma}{numeric. User_defined sizing factor.}

\item{b}{numeric. Size of subsamples.}

\item{s}{numeric. Number of subsamples.}
}
\value{
list of data.frame
}
\description{
Sampling original data into s subsamples without replacement
}
