% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weights}
\alias{weights}
\title{Calculate weights of each observation in subsample
based on multinomial method}
\usage{
weights(subsample, n)
}
\arguments{
\item{subsample}{data.frame}

\item{n}{size of original dataset}
}
\value{
matrix
}
\description{
Calculate weights of each observation in subsample
based on multinomial method
}
\examples{
data(iris)
weights(iris, nrow(iris)*10)
}
